pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "tamagotchai"
version = "1.0.0"
description = "AI service status board for Raspberry Pi e-paper display"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
pisugar = [
    "pisugar>=1.0.0",
]

[tool.setuptools]
packages = ["core", "core.providers", "core.display", "core.screens"]
package-dir = {"." = "."}
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.png", "*.txt"]
